+++ client/src/App.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 // import { AppContextProvider } from './context/AppContext';
-import Homepage from './pages/Homepage';
-import LocalVoice from './pages/LocalVoice';
-import MsgRep from './pages/MsgRep';
+import Homepage from './components/Homepage/Homepage';
+import LocalVoice from './components/LocalVoice/LocalVoice';
+import MsgRep from './components/MsgRep/MsgRep';
 import Navbar from './components/Navbar';
-import ThankYou from './pages/ThankYou';
-import WhoRepYou from './pages/WhoRepYou';
-import YvYv from './pages/YvYv';
-import RepProfile from './pages/RepProfile';
+import ThankYou from './components/ThankYou/ThankYou';
+import WhoRepYou from './components/WhoRepYou/WhoRepYou';
+import YvYv from './components/YvYv/YvYv';
+import RepProfile from './components/RepProfile/RepProfile';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import './App.css';
