+++ client/src/components/Homepage/Homepage.css
@@ -1,0 +1,104 @@
+.body {
+  display: flex-column;
+  position: relative;
+  background-color: blue;
+}
+
+.actiondiv {
+  background-color: navy;
+  color: white;
+  padding: 15px;
+  height: 18rem;
+}
+.actionbutton {
+  background-color: red;
+  box-shadow: 5px 8px black;
+  color: white;
+  border: none;
+  border-radius: 8%;
+  padding: 15px;
+}
+.lhdiv {
+  background-color: lightblue;
+  color: navy;
+  max-width: 350px;
+  border-radius: 5%;
+  box-shadow: 5px 10px;
+  padding: 10px;
+}
+.buttons {
+  display: flex;
+  flex-wrap: wrap;
+  position: relative;
+  align-content: space-around;
+  justify-content: space-evenly;
+}
+.hpconnect {
+  position: relative;
+  color: white;
+  background-color: navy;
+  height: 150px;
+  width: 160px;
+  padding: 10px;
+  box-shadow: 5px 10px navy;
+  border-radius: 5%;
+  margin-top: 15px;
+}
+.hpsubmit {
+  position: relative;
+  color: white;
+  background-color: navy;
+  height: 150px;
+  width: 160px;
+  padding: 10px;
+  box-shadow: 5px 10px navy;
+  border-radius: 5%;
+  margin-top: 15px;
+}
+.hpvoter {
+  position: relative;
+  color: white;
+  background-color: navy;
+  height: 150px;
+  width: 160px;
+  padding: 10px;
+  box-shadow: 5px 10px navy;
+  border-radius: 5%;
+  margin-top: 15px;
+}
+.hpcity {
+  position: relative;
+  color: white;
+  background-color: navy;
+  height: 150px;
+  width: 160px;
+  padding: 10px;
+  box-shadow: 5px 10px darkblue;
+  border-radius: 5%;
+  margin-top: 15px;
+}
+.quiz {
+  position: relative;
+  background-color: lightblue;
+  color: navy;
+  height: 160px;
+  min-width: 300px;
+  max-width: 300px;
+  box-shadow: 5px 10px;
+  border-radius: 5%;
+  padding: 10px;
+  margin-top: 15px;
+}
+.quizbutton {
+  height: 30px;
+  width: 160px;
+  box-shadow: 3px 5px navy;
+  background-color: lightblue;
+  border: 1px solid navy;
+  border-radius: 8%;
+}
+.lhtitle {
+  position: relative;
+  color: white;
+  padding: 10px;
+}
