+++ client/src/components/LocalVoice/LocalVoice.css
@@ -1,0 +1,52 @@
+.body {
+  background-color: blue;
+}
+.lvtitle {
+  background-color: navy;
+  color: white;
+  padding: 10px;
+}
+.title2 {
+  padding: 10px;
+}
+.buttons {
+  display: flex-wrap;
+  justify-content: space-evenly;
+  padding: 10px;
+}
+.findrepbutton {
+  position: relative;
+  min-height: 130px;
+  max-height: 130px;
+  min-width: 350px;
+  max-width: 350px;
+  border-radius: 5%;
+  background-color: lightblue;
+}
+.submitreqbutton {
+  position: relative;
+  min-height: 130px;
+  max-height: 130px;
+  min-width: 350px;
+  max-width: 350px;
+  border-radius: 5%;
+  background-color: lightblue;
+}
+.essentialsbutton {
+  position: relative;
+  min-height: 130px;
+  max-height: 130px;
+  min-width: 350px;
+  max-width: 350px;
+  border-radius: 5%;
+  background-color: lightblue;
+}
+.calendarbutton {
+  position: relative;
+  min-height: 130px;
+  max-height: 130px;
+  min-width: 350px;
+  max-width: 350px;
+  border-radius: 5%;
+  background-color: lightblue;
+}
